Move invokeGuardedCallback into a separate file
This lets us introduce forks for it.
